fix: [ci-fix] revert unrelated database view changes, add page-tree-item stories
- Reverted database view re-consolidation (table-cell, table-row,
  table-column-header, table-skeleton, table-navigation, table-columns,
  filter-value-editor) — these are unrelated to page-tree decomposition
  and should be in a separate PR
- Restored visual regression snapshots for filter-value-editor
- Added page-tree-item.stories.tsx with stories covering default,
  selected, icon, database, untitled, expanded/collapsed, dragged,
  drop targets, favorited, and nested states
