+++ .agents/architecture.md
@@ -287,7 +287,8 @@ Database page (is_database = true)
 src/components/database/
   ├── database-view-client.tsx     # Main client component: loads data, manages view state
   ├── view-tabs.tsx                # Horizontal tab bar for switching views
-  ├── filter-bar.tsx               # Active filter pills + add filter UI
+  ├── filter-bar.tsx               # Active filter pills + add filter UI (composition only)
+  ├── filter-value-editor.tsx      # Type-specific filter value editors, property picker, operator picker
   ├── sort-menu.tsx                # Sort configuration dropdown
   ├── property-config.tsx          # Column header dropdown: rename, type change, delete
   ├── property-editor.tsx          # Inline cell editor (dispatches to type-specific editors)
@@ -309,9 +310,15 @@ src/components/database/
   │   ├── formula.tsx
   │   └── computed.tsx             # created_time, updated_time, created_by (read-only)
   ├── views/
-  │   ├── table-view.tsx           # Spreadsheet grid with resizable columns
-  │   ├── table-cell-renderer.tsx  # CellRenderer + SelectBadge (extracted from table-view)
-  │   ├── table-keyboard.ts       # Keyboard navigation handler (extracted from table-view)
+  │   ├── table-view.tsx           # Composition root — wires sub-components together
+  │   ├── table-row.tsx            # TableRow — title cell + property cells for one row
+  │   ├── table-cell.tsx           # TableCell + RegistryEditorCell — display/edit modes, portal editors
+  │   ├── table-column-header.tsx  # TableColumnHeader — sort, menu, drag, resize handle
+  │   ├── table-skeleton.tsx       # TableSkeleton — loading placeholder
+  │   ├── table-navigation.ts     # useTableCellNavigation — editing/focus state, keyboard nav
+  │   ├── table-columns.ts        # useColumnResize + useColumnDragReorder hooks
+  │   ├── table-cell-renderer.tsx  # CellRenderer + SelectBadge — display-only value rendering
+  │   ├── table-keyboard.ts       # handleCellKeyDown — editing-mode key handler
   │   ├── table-defaults.ts       # Pure helpers: value keys, select options, display values, date formatting
   │   ├── board-view.tsx           # Kanban columns grouped by select property
   │   ├── list-view.tsx            # Compact vertical list
@@ -438,16 +445,23 @@ src/
 │   ├── database/                # Database views system
 │   │   ├── database-view-client.tsx     # Main client component: loads data, manages view state
 │   │   ├── view-tabs.tsx                # Horizontal tab bar for switching views
-│   │   ├── filter-bar.tsx               # Active filter pills + add filter UI
+│   │   ├── filter-bar.tsx               # Active filter pills + add filter UI (composition only)
+│   │   ├── filter-value-editor.tsx      # Type-specific filter value editors, property picker, operator picker
 │   │   ├── sort-menu.tsx                # Sort configuration dropdown
 │   │   ├── property-config.tsx          # Column header dropdown: rename, type change, delete
 │   │   ├── property-editor.tsx          # Inline cell editor (dispatches to type-specific editors)
 │   │   ├── property-renderer.tsx        # Cell renderer (dispatches to type-specific renderers)
 │   │   ├── property-types/              # Registry of type-specific renderers and editors
 │   │   ├── views/
-│   │   │   ├── table-view.tsx           # Spreadsheet grid with resizable columns
-│   │   │   ├── table-cell-renderer.tsx  # CellRenderer + SelectBadge (extracted from table-view)
-│   │   │   ├── table-keyboard.ts       # Keyboard navigation handler (extracted from table-view)
+│   │   │   ├── table-view.tsx           # Composition root — wires sub-components together
+│   │   │   ├── table-row.tsx            # TableRow — title cell + property cells for one row
+│   │   │   ├── table-cell.tsx           # TableCell + RegistryEditorCell — display/edit, portal editors
+│   │   │   ├── table-column-header.tsx  # TableColumnHeader — sort, menu, drag, resize handle
+│   │   │   ├── table-skeleton.tsx       # TableSkeleton — loading placeholder
+│   │   │   ├── table-navigation.ts     # useTableCellNavigation hook — editing/focus/keyboard nav
+│   │   │   ├── table-columns.ts        # useColumnResize + useColumnDragReorder hooks
+│   │   │   ├── table-cell-renderer.tsx  # CellRenderer + SelectBadge — display-only value rendering
+│   │   │   ├── table-keyboard.ts       # handleCellKeyDown — editing-mode key handler
 │   │   │   ├── table-defaults.ts       # Pure helpers: value keys, select options, display values
 │   │   │   ├── board-view.tsx           # Kanban columns grouped by select property
 │   │   │   ├── list-view.tsx            # Compact vertical list
